db_handler/db_handler/util/database_utils.py
import os

from psycopg2 import connect


class DatabaseUtils:
    DATABASE_CONNECTION = connect(
        host=os.environ.get("POSTGRES_HOST"),
        database=os.environ.get("POSTGRES_DB"),
        user=os.environ.get("POSTGRES_USER"),
        password=os.environ.get("POSTGRES_PASSWORD"),
        port=os.environ.get("POSTGRES_PORT")
    )
